major restructuring of project repo
