+++ .github/workflows/continous-deployment.yml
@@ -15,6 +15,9 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v2
 
+      - name: Run cp_shrd.sh
+        run: bash cp_shrd.sh
+
       - name: Login to Docker Hub
         uses: docker/login-action@v1
         with:
