+++ .gitignore
@@ -1,6 +1,10 @@
 .venv/
 __pycache__/
 
+# ignore /utils folder in app and API
+app/*/utils/
+API/*/utils/
+
 # Node Project files to ignore
 node_modules/
