+++ API/package.json
@@ -3,9 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www",
-    "start-api": "node ./API/api.js",
-    "test": "mocha ./tests/*.test.js"
+    "start-api": "node ./src/api.js"
   },
   "dependencies": {
     "bcrypt": "^5.1.1",
@@ -22,13 +20,6 @@
     "os-utils": "^0.0.14",
     "pg": "^8.11.3",
     "prom-client": "^15.1.0",
-    "sqlite3": "^5.1.7",
     "uuid": "^9.0.1"
-  },
-  "devDependencies": {
-    "assert": "^2.1.0",
-    "chai": "^5.1.0",
-    "mocha": "~3.5.3",
-    "supertest": "^6.3.4"
   }
 }
