+++ app/package.json
@@ -3,9 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www",
-    "start-api": "node ./API/api.js",
-    "test": "mocha ./tests/*.test.js"
+    "start": "node ./src/bin/www",
+    "test": "mocha ./src/tests/*.test.js"
   },
   "dependencies": {
     "bcrypt": "^5.1.1",
@@ -22,7 +21,6 @@
     "os-utils": "^0.0.14",
     "pg": "^8.11.3",
     "prom-client": "^15.1.0",
-    "sqlite3": "^5.1.7",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
