+++ app/src/app.js
@@ -12,12 +12,12 @@ var expressLayouts = require('express-ejs-layouts');
 const bcrypt = require('bcrypt');
 const MD5 = require('crypto-js/md5');
 //const { connect_DB, init_DB, query, execute, get_user_id} = require('./utils/dbUtils');
-const { pool, init_DB, query, execute, get_user_id} = require('../../utils/db');
+const { pool, init_DB, query, execute, get_user_id} = require('./utils/db');
 
 
 
 // Prometheus tracking
-const { prometheus, prometheusMiddleware } = require('../../utils/prometheus');
+const { prometheus, prometheusMiddleware } = require('./utils/prometheus');
 
 // Configuration
 const DATABASE = './minitwit.db';
